+++ core/src/com/bonkan/brao/engine/entity/humans/Enemy.java
@@ -7,7 +7,7 @@ import com.bonkan.brao.engine.entity.Human;
 
 public class Enemy extends Human {
 
-	public Enemy(float x, float y, int bodyIndex, int headIndex, UUID id, String userName, World world) {
+	public Enemy(int x, int y, int bodyIndex, int headIndex, UUID id, String userName, World world) {
 		super(x, y, bodyIndex, headIndex, id, userName, world);
 	}
